feat: refactor of seat form into its own component
