+++ lib/types.ts
@@ -42,7 +42,7 @@ export type FlightProps = {
 export type FlightCardProps = {
   flight: FlightProps;
   handleDelete: React.MouseEventHandler<HTMLButtonElement>;
-  handleRemoveFlight: ({flightNumber: string, departureTime:string});
+  handleRemoveFlight: React.MouseEventHandler<HTMLButtonElement>;
   handleUpdateSeat: (seat: Seat, flightNumber:string, oldSeat:string) => void;
   handleUpdatePreferences: (preferences: Preferences,flightNumber: string) => void;
   handleAddFlight: (flight: FlightProps) => void;
